♻️ refactor : 커뮤니티 상세페이지 리엑트마크다운 렌더링부 컴포넌트로 분리
